A0-2236: Fix list item margins
Some of the list items' elements had fixed width in px which
caused the items to overflow on views in which their ancestors
did not have the exact perfect width for which it was designed.
Fixing those elements required changes in several levels
of ancestor elements, turning them into proper flex elements.